requirements.txt
Flask==3.0.3
openai==1.99.9
python-dotenv==1.0.1
azure-search-documents==11.6.0
